+++ test/unit/conftest.py
@@ -69,7 +69,7 @@ def temp_files(tmpdir, request):
 @pytest.fixture()
 def molecule_instance(temp_files, state_path_without_data):
     c = temp_files(fixtures=['molecule_vagrant_config'])
-    m = core.Molecule(dict())
+    m = core.Molecule({})
     m.config = config.Config(configs=c)
     m.state = state.State(state_file=state_path_without_data)
     m.main()