+++ test/unit/driver/test_openstackdriver.py
@@ -33,7 +33,7 @@ from molecule.driver import openstackdriver
 @pytest.fixture()
 def molecule_instance(temp_files, state_path_without_data):
     c = temp_files(fixtures=['molecule_openstack_config'])
-    m = core.Molecule(dict())
+    m = core.Molecule({})
     m.config = config.Config(configs=c)
     m.state = state.State(state_file=state_path_without_data)